chore(types): standardize contract names and types
